Many little changes to get rid of compile-time warnings.
There now should be no warnings when compiling on LINUX. YMMV elsewhere.

+++ connections/canconmanager.cpp
@@ -214,7 +214,7 @@ bool CANConManager::sendFrame(const CANFrame& pFrame)
     foreach (CANConnection* conn, mConns)
     {
         //check if this CAN connection is supposed to handle the requested bus
-        if (pFrame.bus < (uint32_t)(busBase + conn->getNumBuses()))
+        if (pFrame.bus < (busBase + conn->getNumBuses()))
         {
             workingFrame.bus -= busBase;
             workingFrame.isReceived = false;
